backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pdfplumber==0.11.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.6.3
redis==5.0.3
google-genai>=0.8.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
pgvector>=0.3.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
arq>=0.25.0
cryptography>=42.0.0
